feat: reorganize assets for better export
