--- src/assets/styles/components/account.css
@@ -1,155 +1,0 @@
-.account {
-    box-sizing: border-box;
-    width: 100%;
-    height: auto;
-    border-bottom: 1px solid var(--color-light-border);
-}
-
-.account--clickable {
-    cursor: pointer;
-}
-
-.account:last-child {
-    border: none;
-}
-
-.account__section {
-    box-sizing: border-box;
-    width: auto;
-    height: auto;
-    padding-top: 10px;
-    padding-bottom: 10px;
-}
-
-.account__section--right {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-}
-
-.account__accordion {
-    box-sizing: border-box;
-    position: relative;
-    display: block;
-    width: auto;
-}
-
-.account__name {
-    box-sizing: border-box;
-    display: block;
-    font-size: 16px;
-    font-weight: 700;
-    color: var(--color-light-text);
-    line-height: 20px;
-    margin: 0;
-}
-
-.account__name--margin {
-    margin-bottom: 6px;
-}
-
-.account__number {
-    box-sizing: border-box;
-    display: block;
-    margin: 0;
-    font-size: 14px;
-    font-weight: 600;
-    color: var(--color-light-text-alt);
-    line-height: 1;
-}
-
-.account__icon-button {
-    box-sizing: border-box;
-    width: 36px;
-    height: 36px;
-    list-style: none;
-    background-color: transparent;
-    background-image: url("/src/assets/images/delete.svg");
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: 16px;
-    cursor: pointer;
-    margin-right: -10px;
-}
-
-.account__icon-button:hover {
-    background-color: var(--color-light-icon-bg);
-}
-
-.dark .account__icon-button:hover {
-    background-color: var(--color-dark-icon-bg);
-}
-
-.account__text-button {
-    box-sizing: border-box;
-    display: block;
-    position: absolute;
-    top: 0;
-    left: -8px;
-    transform: translateX(-100%);
-    width: auto;
-    height: 36px;
-    padding-left: 16px;
-    padding-right: 16px;
-    border: none;
-    background-color: var(--color-light-down);
-    font-family: inherit;
-    font-size: 10px;
-    font-weight: 700;
-    text-transform: uppercase;
-    letter-spacing: 0.6px;
-    color: #fff;
-    cursor: pointer;
-}
-
-.account__paste-icon {
-    box-sizing: border-box;
-    width: 48px;
-    height: 48px;
-    background-color: transparent;
-    background-image: url("/src/assets/images/paste-form.svg");
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: 24px;
-    margin-right: -12px;
-}
-
-.dark .account__icon {
-    color: var(--color-dark-toggle-button-icon);
-}
-
-.dark .account {
-    border-bottom: 1px solid var(--color-dark-border);
-}
-
-.dark .account:last-child {
-    border: none;
-}
-
-.dark .account__name {
-    color: var(--color-dark-text);
-}
-
-.dark .account__button {
-    background-color: var(--color-dark-down);
-}
-
-.dark .account__number {
-    color: var(--color-dark-text-alt);
-}
-
-.maze .account--clickable:hover .account__name {
-    color: var(--color-light-maze);
-}
-
-.fleeca .account--clickable:hover .account__name {
-    color: var(--color-light-fleeca);
-}
-
-.maze.dark .account--clickable:hover .account__name {
-    color: var(--color-dark-maze);
-}
-
-.fleeca.dark .account--clickable:hover .account__name {
-    color: var(--color-dark-fleeca);
-}
