--- src/assets/styles/components/app.css
@@ -1,75 +1,0 @@
-@keyframes spin {
-    0% {
-        transform: rotate(0deg);
-    }
-
-    100% {
-        transform: rotate(360deg);
-    }
-}
-
-.app {
-    box-sizing: border-box;
-    display: flex;
-    width: 1224px;
-    height: 720px;
-    margin: auto;
-    background-color: var(--color-light-bg);
-    border: 1px solid var(--color-light-border-on-bg);
-    overflow: hidden;
-}
-
-.app__spinner {
-    display: none;
-    box-sizing: border-box;
-    width: 72px;
-    height: 72px;
-    border-radius: 9999px;
-    border: 8px solid var(--color-light-toggle-button-track);
-    animation: spin 2s linear infinite;
-    margin: auto;
-}
-
-.app__container {
-    box-sizing: border-box;
-    display: none;
-    width: 100%;
-    height: 100%;
-}
-
-.app__tab-selector {
-    position: absolute;
-    width: 1px;
-    height: 1px;
-    padding: 0;
-    margin: -1px;
-    overflow: hidden;
-    clip: rect(0, 0, 0, 0);
-    white-space: nowrap;
-    border-width: 0;
-}
-
-.dark .app {
-    background-color: var(--color-dark-bg);
-    border: 1px solid var(--color-dark-border-on-bg);
-}
-
-.maze .app__spinner {
-    border-top: 8px solid var(--color-light-maze);
-}
-
-.fleeca .app__spinner {
-    border-top: 8px solid var(--color-light-fleeca);
-}
-
-.dark .app__spinner {
-    border: 8px solid var(--color-dark-toggle-button-track);
-}
-
-.maze.dark .app__spinner {
-    border-top: 8px solid var(--color-dark-maze);
-}
-
-.fleeca.dark .app__spinner {
-    border-top: 8px solid var(--color-dark-fleeca);
-}
