--- src/assets/styles/components/balance.css
@@ -1,64 +1,0 @@
-.balance {
-    box-sizing: border-box;
-    width: auto;
-    height: auto;
-}
-
-.balance__name {
-    font-size: 10px;
-    font-weight: 700;
-    letter-spacing: 0.4px;
-    color: var(--color-light-text);
-    text-transform: uppercase;
-    line-height: 10px;
-    margin: 0 0 4px 0;
-}
-
-.balance__value {
-    display: inline;
-    font-size: 24px;
-    font-weight: 700;
-    letter-spacing: 0.8px;
-    color: transparent;
-    text-shadow: 0 0 12px var(--color-light-up);
-    text-transform: uppercase;
-    line-height: 24px;
-    margin: 0;
-    cursor: default;
-}
-
-.balance__value:hover {
-    color: var(--color-light-up);
-    text-shadow: none;
-}
-
-.balance__value--negative {
-    text-shadow: 0 0 12px var(--color-light-down);
-}
-
-.balance__value--negative:hover {
-    color: var(--color-light-down);
-    text-shadow: none;
-}
-
-.dark .balance__name {
-    color: var(--color-dark-text);
-}
-
-.dark .balance__value {
-    text-shadow: 0 0 12px var(--color-dark-up);
-}
-
-.dark .balance__value:hover {
-    color: var(--color-dark-up);
-    text-shadow: none;
-}
-
-.dark .balance__value--negative {
-    text-shadow: 0 0 12px var(--color-dark-down);
-}
-
-.dark .balance__value--negative:hover {
-    color: var(--color-dark-down);
-    text-shadow: none;
-}
