--- src/assets/styles/components/bar.css
@@ -1,74 +1,0 @@
-.bar {
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
-    height: 72px;
-    padding-left: 24px;
-    padding-right: 24px;
-    background-color: var(--color-light-fg);
-    border-bottom: 0.1px solid var(--color-light-border);
-}
-
-.bar__section {
-    box-sizing: border-box;
-    flex: 1 1 0%;
-}
-
-.bar__section--middle {
-    height: 100%;
-}
-
-.bar__section--right {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    align-items: flex-start;
-}
-
-.bar__section--row {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-    align-items: center;
-}
-
-.bar__logo {
-    box-sizing: border-box;
-    display: block;
-    width: 36px;
-    height: auto;
-    margin-right: 16px;
-    pointer-events: none;
-}
-
-.dark .bar {
-    background-color: var(--color-dark-fg);
-    border-bottom: 0.1px solid var(--color-dark-border);
-}
-
-.maze .bar__logo {
-    color: var(--color-light-maze);
-}
-
-.maze.dark .bar__logo {
-    color: var(--color-dark-maze);
-}
-
-.fleeca .bar__logo {
-    color: var(--color-light-fleeca);
-}
-
-.fleeca.dark .bar__logo {
-    color: var(--color-dark-fleeca);
-}
-
-.maze .bar__logo--fleeca {
-    display: none;
-}
-
-.fleeca .bar__logo--maze {
-    display: none;
-}
