--- src/assets/styles/components/button.css
@@ -1,77 +1,0 @@
-.button {
-    box-sizing: border-box;
-    height: 48px;
-    cursor: pointer;
-    background-color: transparent;
-    border: none;
-}
-
-.button--text {
-    display: block;
-    min-width: 168px;
-    padding-left: 24px;
-    padding-right: 24px;
-    font-family: inherit;
-    font-size: 12px;
-    font-weight: 700;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-}
-
-.button--icon {
-    display: flex;
-    width: 48px;
-}
-
-.button__icon {
-    box-sizing: border-box;
-    width: 24px;
-    height: 24px;
-    margin: auto;
-}
-
-.button--primary {
-    color: #ffffff;
-}
-
-.maze .button--primary {
-    background-color: var(--color-light-maze);
-}
-
-.fleeca .button--primary {
-    background-color: var(--color-light-fleeca);
-}
-
-.maze.dark .button--primary {
-    background-color: var(--color-dark-maze);
-}
-
-.fleeca.dark .button--primary {
-    background-color: var(--color-dark-fleeca);
-}
-
-.maze .button--secondary {
-    color: var(--color-light-maze);
-}
-
-.fleeca .button--secondary {
-    color: var(--color-light-fleeca);
-}
-
-.maze.dark .button--secondary {
-    color: var(--color-dark-maze);
-}
-
-.fleeca.dark .button--secondary {
-    color: var(--color-dark-fleeca);
-}
-
-.button--disabled {
-    color: var(--color-light-text-alt);
-    background-color: #ffffff;
-}
-
-.dark .button--disabled {
-    color: var(--color-dark-text-alt);
-    background-color: var(--color-dark-icon-bg);
-}
