--- src/assets/styles/components/column.css
@@ -1,23 +1,0 @@
-.column {
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    height: 100%;
-}
-
-.column__section {
-    box-sizing: border-box;
-    width: 100%;
-    height: auto;
-}
-
-.column__section--fixed {
-    flex: 0 0 auto;
-}
-
-.column__section--fluid {
-    flex: 1 1 0%;
-}
-
-
