--- src/assets/styles/components/form.css
@@ -1,123 +1,0 @@
-.form {
-    box-sizing: border-box;
-}
-
-.form__title {
-    box-sizing: border-box;
-    font-size: 18px;
-    font-weight: 700;
-    color: var(--color-light-text);
-    line-height: 1;
-    margin: 0 0 20px 0;
-}
-
-.form__row {
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: row;
-    gap: 24px;
-}
-
-.form__row--margin {
-    margin-bottom: 24px;
-}
-
-.form__row--items-end {
-    align-items: flex-end;
-}
-
-.form__section {
-    box-sizing: border-box;
-    position: relative;
-}
-
-.form__section--margin {
-    margin-bottom: 24px;
-}
-
-.form__row > .form__section--fluid {
-    flex: 1 1 0%;
-}
-
-.form__row > .form__section--fixed {
-    flex: 0 0 auto;
-}
-
-.form__label {
-    box-sizing: border-box;
-    display: block;
-    font-size: 12px;
-    font-weight: 600;
-    text-transform: uppercase;
-    letter-spacing: 0.6px;
-    color: var(--color-light-text-alt);
-    line-height: 1;
-    margin-bottom: 8px;
-}
-
-.form__input {
-    box-sizing: border-box;
-    display: block;
-    width: 100%;
-    height: 48px;
-    margin: 0;
-    padding: 0 16px;
-    border: 1px solid var(--color-light-border);
-    color: var(--color-light-text);
-    background-color: var(--color-light-fg);
-    font-family: inherit;
-    font-size: 14px;
-    font-weight: 700;
-}
-
-.form__input::placeholder {
-    font-weight: 600;
-    color: var(--color-light-text-placeholder);
-}
-
-.form__input:focus {
-    outline: none;
-}
-
-.form__message {
-    position: absolute;
-    margin: 2px 0 0 0;
-    font-size: 12px;
-    font-weight: 700;
-    color: var(--color-light-down);
-    line-height: 16px;
-}
-
-.maze .form__input:focus {
-    border: 1px solid var(--color-light-maze);
-}
-
-.fleeca .form__input:focus {
-    border: 1px solid var(--color-light-fleeca);
-}
-
-.maze.dark .form__input:focus {
-    border: 1px solid var(--color-dark-maze);
-}
-
-.fleeca.dark .form__input:focus {
-    border: 1px solid var(--color-dark-fleeca);
-}
-
-.dark .form__title {
-    color: var(--color-dark-text);
-}
-
-.dark .form__label {
-    color: var(--color-dark-text-alt);
-}
-
-.dark .form__input {
-    border: 1px solid var(--color-dark-border);
-    color: var(--color-dark-text);
-    background-color: var(--color-dark-fg);
-}
-
-.dark .form__input::placeholder {
-    color: var(--color-dark-text-placeholder);
-}
