--- src/assets/styles/components/log.css
@@ -1,171 +1,0 @@
-.log {
-    box-sizing: border-box;
-    width: 100%;
-    height: auto;
-    padding-top: 16px;
-    padding-bottom: 16px;
-    border-bottom: 1px solid var(--color-light-border);
-}
-
-.log:last-child {
-    border: none;
-}
-
-.log__header {
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: row;
-    align-items: flex-start;
-    width: 100%;
-    height: auto;
-    cursor: pointer;
-    list-style: none;
-}
-
-.log__body {
-    box-sizing: border-box;
-    padding-top: 8px;
-}
-
-.log__icon {
-    flex: 0 0 auto;
-    box-sizing: border-box;
-    width: 48px;
-    height: 48px;
-    margin-right: 16px;
-    background-color: var(--color-light-icon-bg);
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: 24px;
-}
-
-.log__icon--person {
-    background-image: url("/src/assets/images/person.svg");
-}
-
-.log__icon--police {
-    background-image: url("/src/assets/images/police.svg");
-}
-
-.log__icon--bank {
-    background-image: url("/src/assets/images/bank.svg");
-}
-
-.log__amount {
-    flex: 0 0 auto;
-}
-
-.log__info {
-    flex: 1 1 0%;
-    margin-top: 5px;
-    margin-bottom: 5px;
-}
-
-.log__type {
-    box-sizing: border-box;
-    display: block;
-    font-size: 16px;
-    font-weight: 700;
-    color: var(--color-light-text);
-    line-height: 1;
-    margin: 0 0 8px 0;
-}
-
-.log__date {
-    box-sizing: border-box;
-    display: block;
-    margin: 0;
-    font-size: 14px;
-    font-weight: 600;
-    color: var(--color-light-text-alt);
-    line-height: 1;
-}
-
-.log__amount {
-    box-sizing: border-box;
-    display: block;
-    font-size: 16px;
-    font-weight: 700;
-    line-height: 1;
-    margin: 5px 0 0 0;
-}
-
-.log__amount--up {
-    color: var(--color-light-up);
-}
-
-.log__amount--down {
-    color: var(--color-light-down);
-}
-
-.log__reference-label {
-    box-sizing: border-box;
-    display: inline;
-    margin: 0 0 8px 0;
-    font-size: 14px;
-    font-weight: 700;
-    color: var(--color-light-text);
-    line-height: 20px;
-}
-
-.log__reference {
-    box-sizing: border-box;
-    display: inline;
-    margin: 0;
-    font-size: 14px;
-    font-weight: 600;
-    color: var(--color-light-text-alt);
-    line-height: 20px;
-}
-
-.dark .log {
-    border-bottom: 1px solid var(--color-dark-border);
-}
-
-.dark .log:last-child {
-    border: none;
-}
-
-.dark .log__icon {
-    background-color: var(--color-dark-icon-bg);
-}
-
-.dark .log__type {
-    color: var(--color-dark-text);
-}
-
-.dark .log__date {
-    color: var(--color-dark-text-alt);
-}
-
-.dark .log__reference-label {
-    color: var(--color-dark-text);
-}
-
-.dark .log__reference {
-    color: var(--color-dark-text-alt);
-}
-
-.dark .log__amount--up {
-    color: var(--color-dark-up);
-}
-
-.dark .log__amount--down {
-    color: var(--color-dark-down);
-}
-
-.maze .log__header:hover .log__type {
-    color: var(--color-light-maze);
-}
-
-.fleeca .log__header:hover .log__type {
-    color: var(--color-light-fleeca);
-}
-
-.maze.dark .log__header:hover .log__type {
-    color: var(--color-dark-maze);
-}
-
-.fleeca.dark .log__header:hover .log__type {
-    color: var(--color-dark-fleeca);
-}
