--- src/assets/styles/components/screen.css
@@ -1,12 +1,0 @@
-.screen {
-    box-sizing: border-box;
-    display: flex;
-    width: 100%;
-    min-height: 100vh;
-    margin: 0;
-    background-color: var(--color-light-fg);
-}
-
-.dark .screen {
-    background-color: #1A1B1E;
-}
