--- src/assets/styles/components/tab-list.css
@@ -1,36 +1,0 @@
-.tab-list {
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: row;
-    width: auto;
-    height: 100%;
-}
-
-.tab-list__item {
-    box-sizing: border-box;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    width: 124px;
-    height: 100%;
-    border-top: 4px solid transparent;
-    border-bottom: 4px solid transparent;
-    cursor: pointer;
-}
-
-.tab-list__item--hidden {
-    display: none;
-}
-
-.tab-list__label {
-    font-size: 12px;
-    font-weight: 700;
-    text-transform: uppercase;
-    letter-spacing: 0.6px;
-    color: var(--color-light-text-alt);
-}
-
-.dark .tab-list__label {
-    color: var(--color-dark-text-alt);
-}
