--- src/assets/styles/components/toggle-button.css
@@ -1,91 +1,0 @@
-.toggle-button {
-    box-sizing: border-box;
-    position: relative;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    width: auto;
-    height: auto;
-    padding-top: 16px;
-    padding-bottom: 16px;
-}
-
-.toggle-button__input {
-    box-sizing: border-box;
-    display: block;
-    position: absolute;
-    z-index: 20;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    -webkit-appearance: none;
-    appearance: none;
-    cursor: pointer;
-}
-
-.toggle-button__icon {
-    box-sizing: border-box;
-    width: 16px;
-    height: auto;
-}
-
-.toggle-button__icon--sun {
-    color: var(--color-light-toggle-button-icon-active);
-}
-
-.toggle-button__icon--moon {
-    color: var(--color-light-toggle-button-icon);
-}
-
-.dark .toggle-button__icon--sun {
-    color: var(--color-dark-toggle-button-icon);
-}
-
-.dark .toggle-button__icon--moon {
-    color: var(--color-dark-toggle-button-icon-active);
-}
-
-.toggle-button__track {
-    box-sizing: border-box;
-    width: 40px;
-    height: 20px;
-    padding: 4px;
-    margin-left: 12px;
-    margin-right: 12px;
-    border-radius: 9999px;
-    background-color: var(--color-light-toggle-button-track);
-}
-
-.toggle-button__handle {
-    box-sizing: border-box;
-    position: relative;
-    z-index: 10;
-    width: 12px;
-    height: 12px;
-    border-radius: 9999px;
-}
-
-.maze .toggle-button__handle {
-    background-color: var(--color-light-maze);
-}
-
-.fleeca .toggle-button__handle {
-    background-color: var(--color-light-fleeca);
-}
-
-.toggle-button__input:checked ~ .toggle-button__track > .toggle-button__handle {
-    transform: translateX(20px);
-}
-
-.dark .toggle-button__track {
-    background-color: var(--color-dark-toggle-button-track);
-}
-
-.maze.dark .toggle-button__handle {
-    background-color: var(--color-dark-maze);
-}
-
-.fleeca.dark .toggle-button__handle {
-    background-color: var(--color-dark-fleeca);
-}
