--- src/assets/styles/components/view.css
@@ -1,7 +1,0 @@
-.view {
-    box-sizing: border-box;
-    display: none;
-    width: 100%;
-    height: calc(100% - 72px);
-    padding: 24px;
-}
