--- src/assets/styles/global.css
@@ -1,45 +1,0 @@
-:root {
-    --color-dark-screen: #1A1B1E;
-    --color-dark-bg: #121417;
-    --color-dark-fg: #191C1F;
-    --color-dark-text: #CED4DA;
-    --color-dark-text-alt: #868E96;
-    --color-dark-text-placeholder: #5D666F;
-    --color-dark-border: #22262A;
-    --color-dark-border-on-bg: #22262A;
-    --color-dark-scrollbar-track: #1A1B1E;
-    --color-dark-scrollbar-thumb: #272B30;
-    --color-dark-icon-bg: #272B30;
-    --color-dark-button-secondary: #272B30;
-    --color-dark-toggle-button-track: #383D43;
-    --color-dark-toggle-button-icon: #868E96;
-    --color-dark-toggle-button-icon-active: #fcc419;
-    --color-dark-up: #0ca678;
-    --color-dark-down: #d6336c;
-    --color-dark-fleeca: #37b24d;
-    --color-dark-maze: #c92a2a;
-
-    --color-light-screen: #e9ecef;
-    --color-light-bg: #F3F5F6;
-    --color-light-fg: #ffffff;
-    --color-light-text: #495057;
-    --color-light-text-alt: #868e96;
-    --color-light-text-placeholder: #adb5bd;
-    --color-light-border: #e9ecef;
-    --color-light-border-on-bg: #dee2e6;
-    --color-light-scrollbar-track: #ced4da;
-    --color-light-scrollbar-thumb: #dee2e6;
-    --color-light-icon-bg: #f1f3f5;
-    --color-light-button-secondary: #e9ecef;
-    --color-light-toggle-button-track: #e9ecef;
-    --color-light-toggle-button-icon: #9299A0;
-    --color-light-toggle-button-icon-active: #f59f00;
-    --color-light-up: #20c997;
-    --color-light-down: #f06595;
-    --color-light-fleeca: #40c057;
-    --color-light-maze: #ff6b6b;
-}
-
-html {
-    font-family: "Red Hat Display", sans-serif;
-}
