--- src/assets/styles/states.css
@@ -1,47 +1,0 @@
-.maze #tab-selector-account:checked ~ .bar #tab-account {border-bottom: 4px solid var(--color-light-maze);}
-.maze #tab-selector-operation:checked ~ .bar #tab-operation {border-bottom: 4px solid var(--color-light-maze);}
-.maze #tab-selector-transfer:checked ~ .bar #tab-transfer {border-bottom: 4px solid var(--color-light-maze);}
-.maze #tab-selector-enterprise:checked ~ .bar #tab-enterprise {border-bottom: 4px solid var(--color-light-maze);}
-
-.fleeca #tab-selector-account:checked ~ .bar #tab-account {border-bottom: 4px solid var(--color-light-fleeca);}
-.fleeca #tab-selector-operation:checked ~ .bar #tab-operation {border-bottom: 4px solid var(--color-light-fleeca);}
-.fleeca #tab-selector-transfer:checked ~ .bar #tab-transfer {border-bottom: 4px solid var(--color-light-fleeca);}
-.fleeca #tab-selector-enterprise:checked ~ .bar #tab-enterprise {border-bottom: 4px solid var(--color-light-fleeca);}
-
-.maze.dark #tab-selector-account:checked ~ .bar #tab-account {border-bottom: 4px solid var(--color-dark-maze);}
-.maze.dark #tab-selector-operation:checked ~ .bar #tab-operation {border-bottom: 4px solid var(--color-dark-maze);}
-.maze.dark #tab-selector-transfer:checked ~ .bar #tab-transfer {border-bottom: 4px solid var(--color-dark-maze);}
-.maze.dark #tab-selector-enterprise:checked ~ .bar #tab-enterprise {border-bottom: 4px solid var(--color-dark-maze);}
-
-.fleeca.dark #tab-selector-account:checked ~ .bar #tab-account {border-bottom: 4px solid var(--color-dark-fleeca);}
-.fleeca.dark #tab-selector-operation:checked ~ .bar #tab-operation {border-bottom: 4px solid var(--color-dark-fleeca);}
-.fleeca.dark #tab-selector-transfer:checked ~ .bar #tab-transfer {border-bottom: 4px solid var(--color-dark-fleeca);}
-.fleeca.dark #tab-selector-enterprise:checked ~ .bar #tab-enterprise {border-bottom: 4px solid var(--color-dark-fleeca);}
-
-#tab-selector-account:checked ~ .bar #tab-account > .tab-list__label {color: var(--color-light-text);}
-#tab-selector-operation:checked ~ .bar #tab-operation > .tab-list__label {color: var(--color-light-text);}
-#tab-selector-transfer:checked ~ .bar #tab-transfer > .tab-list__label {color: var(--color-light-text);}
-#tab-selector-enterprise:checked ~ .bar #tab-enterprise > .tab-list__label {color: var(--color-light-text);}
-
-.dark #tab-selector-account:checked ~ .bar #tab-account > .tab-list__label {color: var(--color-dark-text);}
-.dark #tab-selector-operation:checked ~ .bar #tab-operation > .tab-list__label {color: var(--color-dark-text);}
-.dark #tab-selector-transfer:checked ~ .bar #tab-transfer > .tab-list__label {color: var(--color-dark-text);}
-.dark #tab-selector-enterprise:checked ~ .bar #tab-enterprise > .tab-list__label {color: var(--color-dark-text);}
-
-#tab-selector-account:checked ~ #tab-content-account {display: block;}
-#tab-selector-operation:checked ~ #tab-content-operation {display: block;}
-#tab-selector-transfer:checked ~ #tab-content-transfer {display: block;}
-#tab-selector-enterprise:checked ~ #tab-content-enterprise {display: block;}
-#tab-selector-offshore:checked ~ #tab-content-offshore {display: block;}
-
-.app--loading > .app__spinner {
-    display: block;
-}
-
-.app--loaded > .app__container {
-    display: block;
-}
-
-.account__accordion[open] > .account__icon-button {
-    background-image: url("/src/assets/images/close.svg");
-}
