+++ src/data.js
@@ -1,0 +1,195 @@
+const data = {
+  bank: "fleeca",
+  darkMode: true,
+  name: "Hubert Bonisseur de La Bath",
+  balance: 2152800,
+  number: "0123456789",
+  accounts: [
+    {
+      name: "Lena Silvo",
+      code: "0123456789",
+    },
+    {
+      name: "Owen Chapman",
+      code: "1234567890",
+    },
+    {
+      name: "Garry Green",
+      code: "2345678901",
+    },
+    {
+      name: "Arthur Popov",
+      code: "3456789012",
+    },
+    {
+      name: "Quentin Cooper",
+      code: "4567890123",
+    },
+  ],
+  logs: [
+    {
+      entity: "Los Santos Police Department",
+      date: "2022-04-20T15:00:00",
+      amount: -2455,
+      reference: "Amende pour refus d'optempérer",
+      type: "fine",
+      icon: "police",
+    },
+    {
+      entity: "Transfert",
+      date: "2022-04-19T15:00",
+      amount: 2580,
+      reference: "Argent pour les qualifications des Failygames",
+      type: "transfer",
+      icon: "person",
+    },
+    {
+      entity: "Dépot",
+      date: "2022-04-18T15:00:00",
+      amount: 15000,
+      reference: "Vente de la voiture",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Transfert",
+      date: "2022-04-17T15:00:00",
+      amount: 20000,
+      reference: "Argent pour les qualifications des Failygames",
+      type: "transfer",
+      icon: "person",
+    },
+    {
+      entity: "Retrait",
+      date: "2022-04-16T15:00:00",
+      amount: -4500,
+      reference: "Achat ARC",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Los Santos Police Department",
+      date: "2022-04-15T15:00:00",
+      amount: -100000,
+      reference: "Amende pour braquage de banque",
+      type: "fine",
+      icon: "police",
+    },
+    {
+      entity: "Dépot",
+      date: "2022-04-14T15:00:00",
+      amount: 1200,
+      reference: "Vente de la voiture",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Retrait",
+      date: "2022-04-13T15:00:00",
+      amount: -8500,
+      reference: "Achat ARC",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Transfert",
+      date: "2022-04-12T15:00:00",
+      amount: 15000,
+      reference: "Argent pour les qualifications des Failygames",
+      type: "transfer",
+      icon: "person",
+    },
+    {
+      entity: "Los Santos Police Department",
+      date: "2022-04-20T15:00:00",
+      amount: -2455,
+      reference: "Amende pour refus d'optempérer",
+      type: "fine",
+      icon: "police",
+    },
+    {
+      entity: "Transfert",
+      date: "2022-04-19T15:00",
+      amount: 2580,
+      reference: "Argent pour les qualifications des Failygames",
+      type: "transfer",
+      icon: "person",
+    },
+    {
+      entity: "Dépot",
+      date: "2022-04-18T15:00:00",
+      amount: 15000,
+      reference: "Vente de la voiture",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Transfert",
+      date: "2022-04-17T15:00:00",
+      amount: 20000,
+      reference: "Argent pour les qualifications des Failygames",
+      type: "transfer",
+      icon: "person",
+    },
+    {
+      entity: "Retrait",
+      date: "2022-04-16T15:00:00",
+      amount: -4500,
+      reference: "Achat ARC",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Los Santos Police Department",
+      date: "2022-04-15T15:00:00",
+      amount: -100000,
+      reference: "Amende pour braquage de banque",
+      type: "fine",
+      icon: "police",
+    },
+    {
+      entity: "Dépot",
+      date: "2022-04-14T15:00:00",
+      amount: 1200,
+      reference: "Vente de la voiture",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Retrait",
+      date: "2022-04-13T15:00:00",
+      amount: -8500,
+      reference: "Achat ARC",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Transfert",
+      date: "2022-04-12T15:00:00",
+      amount: 15000,
+      reference: "Argent pour les qualifications des Failygames",
+      type: "transfer",
+      icon: "person",
+    },
+    {
+      entity: "Retrait",
+      date: "2022-04-13T15:00:00",
+      amount: -8500,
+      reference: "Achat ARC",
+      type: "operation",
+      icon: "bank",
+    },
+    {
+      entity: "Transfert",
+      date: "2022-04-12T15:00:00",
+      amount: 15000,
+      reference: "Argent pour les qualifications des Failygames",
+      type: "transfer",
+      icon: "person",
+    },
+  ],
+  hasEnterprise: true,
+  hasOffShore: true,
+};
+
+export default data;
