+++ src/main.js
@@ -1,25 +1,5 @@
-// CSS / Global
-import "./assets/styles/global.css";
-
-// CSS / Components
-import "./assets/styles/components/screen.css";
-import "./assets/styles/components/app.css";
-import "./assets/styles/components/bar.css";
-import "./assets/styles/components/balance.css";
-import "./assets/styles/components/tab-list.css";
-import "./assets/styles/components/toggle-button.css";
-import "./assets/styles/components/view.css";
-import "./assets/styles/components/grid.css";
-import "./assets/styles/components/column.css";
-import "./assets/styles/components/paper.css";
-import "./assets/styles/components/form.css";
-import "./assets/styles/components/section.css";
-import "./assets/styles/components/account.css";
-import "./assets/styles/components/log.css";
-import "./assets/styles/components/button.css";
-
-// CSS / States
-import "./assets/styles/states.css";
+import "./main.css";
+import data from "./data";
 
 // Account elements
 const accountName = document.getElementById("account-name");
@@ -89,203 +69,6 @@ const dateTimeFormatter = new Intl.DateTimeFormat("fr-FR", {
 });
 
 /**
- * POVERS: remove this variable when finish to plug init function in your script.
- */
-const data = {
-    bank: "fleeca",
-    darkMode: true,
-    name: "Hubert Bonisseur de La Bath",
-    balance: 2152800,
-    number: "0123456789",
-    accounts: [
-        {
-            name: "Lena Silvo",
-            code: "0123456789",
-        },
-        {
-            name: "Owen Chapman",
-            code: "1234567890",
-        },
-        {
-            name: "Garry Green",
-            code: "2345678901",
-        },
-        {
-            name: "Arthur Popov",
-            code: "3456789012",
-        },
-        {
-            name: "Quentin Cooper",
-            code: "4567890123",
-        },
-    ],
-    logs: [
-        {
-            entity: "Los Santos Police Department",
-            date: "2022-04-20T15:00:00",
-            amount: -2455,
-            reference: "Amende pour refus d'optempérer",
-            type: "fine",
-            icon: "police",
-        },
-        {
-            entity: "Transfert",
-            date: "2022-04-19T15:00",
-            amount: 2580,
-            reference: "Argent pour les qualifications des Failygames",
-            type: "transfer",
-            icon: "person",
-        },
-        {
-            entity: "Dépot",
-            date: "2022-04-18T15:00:00",
-            amount: 15000,
-            reference: "Vente de la voiture",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Transfert",
-            date: "2022-04-17T15:00:00",
-            amount: 20000,
-            reference: "Argent pour les qualifications des Failygames",
-            type: "transfer",
-            icon: "person",
-        },
-        {
-            entity: "Retrait",
-            date: "2022-04-16T15:00:00",
-            amount: -4500,
-            reference: "Achat ARC",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Los Santos Police Department",
-            date: "2022-04-15T15:00:00",
-            amount: -100000,
-            reference: "Amende pour braquage de banque",
-            type: "fine",
-            icon: "police",
-        },
-        {
-            entity: "Dépot",
-            date: "2022-04-14T15:00:00",
-            amount: 1200,
-            reference: "Vente de la voiture",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Retrait",
-            date: "2022-04-13T15:00:00",
-            amount: -8500,
-            reference: "Achat ARC",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Transfert",
-            date: "2022-04-12T15:00:00",
-            amount: 15000,
-            reference: "Argent pour les qualifications des Failygames",
-            type: "transfer",
-            icon: "person",
-        },
-        {
-            entity: "Los Santos Police Department",
-            date: "2022-04-20T15:00:00",
-            amount: -2455,
-            reference: "Amende pour refus d'optempérer",
-            type: "fine",
-            icon: "police",
-        },
-        {
-            entity: "Transfert",
-            date: "2022-04-19T15:00",
-            amount: 2580,
-            reference: "Argent pour les qualifications des Failygames",
-            type: "transfer",
-            icon: "person",
-        },
-        {
-            entity: "Dépot",
-            date: "2022-04-18T15:00:00",
-            amount: 15000,
-            reference: "Vente de la voiture",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Transfert",
-            date: "2022-04-17T15:00:00",
-            amount: 20000,
-            reference: "Argent pour les qualifications des Failygames",
-            type: "transfer",
-            icon: "person",
-        },
-        {
-            entity: "Retrait",
-            date: "2022-04-16T15:00:00",
-            amount: -4500,
-            reference: "Achat ARC",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Los Santos Police Department",
-            date: "2022-04-15T15:00:00",
-            amount: -100000,
-            reference: "Amende pour braquage de banque",
-            type: "fine",
-            icon: "police",
-        },
-        {
-            entity: "Dépot",
-            date: "2022-04-14T15:00:00",
-            amount: 1200,
-            reference: "Vente de la voiture",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Retrait",
-            date: "2022-04-13T15:00:00",
-            amount: -8500,
-            reference: "Achat ARC",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Transfert",
-            date: "2022-04-12T15:00:00",
-            amount: 15000,
-            reference: "Argent pour les qualifications des Failygames",
-            type: "transfer",
-            icon: "person",
-        },
-        {
-            entity: "Retrait",
-            date: "2022-04-13T15:00:00",
-            amount: -8500,
-            reference: "Achat ARC",
-            type: "operation",
-            icon: "bank",
-        },
-        {
-            entity: "Transfert",
-            date: "2022-04-12T15:00:00",
-            amount: 15000,
-            reference: "Argent pour les qualifications des Failygames",
-            type: "transfer",
-            icon: "person",
-        },
-    ],
-    hasEnterprise: true,
-    hasOffShore: true,
-};
-
-/**
  * Current state of the account.
  */
 let activeAccount = {};
@@ -303,8 +86,7 @@ function displayBalance(newBalance) {
             balanceText.classList.remove("balance__value--negative");
         }
     }
-    const formatedBalance = numberFormatter.format(newBalance);
-    balanceText.textContent = formatedBalance;
+    balanceText.textContent = numberFormatter.format(newBalance);
 }
 
 /**
